LG-2501: Refresh toggle (#1458)
* add darkmode

* changeset

* feedback
