+++ .changeset/pretty-comics-act.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/toggle': major
+---
+
+Updates `Toggle` for dark mode brand refresh
