+++ packages/toggle/package.json
@@ -23,9 +23,12 @@
     "@leafygreen-ui/a11y": "^1.3.3",
     "@leafygreen-ui/emotion": "^4.0.3",
     "@leafygreen-ui/icon": "^11.11.1",
-    "@leafygreen-ui/interaction-ring": "^4.0.3",
     "@leafygreen-ui/lib": "^9.5.1",
-    "@leafygreen-ui/palette": "^3.4.4"
+    "@leafygreen-ui/palette": "^3.4.4",
+    "@leafygreen-ui/tokens": "^1.3.4"
+  },
+  "peerDependencies": {
+    "@leafygreen-ui/leafygreen-provider":"^2.3.5"
   },
   "gitHead": "dd71a2d404218ccec2e657df9c0263dc1c15b9e0",
   "homepage": "https://github.com/mongodb/leafygreen-ui/tree/main/packages/toggle",
