+++ packages/toggle/src/Toggle.story.tsx
@@ -3,6 +3,7 @@ import { ComponentStory } from '@storybook/react';
 import { css } from '@leafygreen-ui/emotion';
 import { palette } from '@leafygreen-ui/palette';
 import Toggle from '.';
+import { Size } from './types';
 
 export default {
   title: 'Components/Toggle',
@@ -13,6 +14,18 @@ export default {
     },
     checked: {
       control: 'boolean',
+    },
+    disabled: {
+      control: 'boolean',
+    },
+    size: {
+      control: 'radio',
+      options: Object.values(Size),
+    },
+  },
+  parameters: {
+    controls: {
+      exclude: ['className', 'onChange'],
     },
   },
 };
