+++ packages/toggle/src/index.ts
@@ -1,3 +1,4 @@
 import Toggle from './Toggle';
-export { Size } from './Toggle';
+export { Size } from './types';
+export type { ToggleProps } from './types';
 export default Toggle;
